Added function to calculate integral in the gate.  Has been tested in the simulator.  Currently hardcoded to call this function instead of the peak find function and to attenuate the result by shifting right 16 bits -- this is most likely too much attenuation.
Fixed comments, mostly related to "gate exceeded" and "gate not exceeded"
variables and clarifying that the gate width value is 1/3 the actual width.

Repaired bug in checkForNew peak where A was not actually being loaded
with the value in AR4 but loading from a memory location in the DP page.
Fixed by using ldm instead of ld.

Signed-off-by: Mike <mike@folsomvillage.com>