Added constant to set the size of the gate parameter, result and DAC buffers so that the buffer size can be increased and new values added at the end of each gate without having to rewrite a lot of code.
Added a value for signal processing named Threshold1 to the gate parameter
buffer.  This can be used in different ways by the various signal
processing functions.

Added stub for catching bad interface (see GATE_QUENCH_ON_OVERLIMIT). If
a gate has this flag set, it will integrate above the gate -- BUT in the
current code the value is ignored -- need to add code to quench.

Added code to calculateGateIntegral to skip the integration if the gate
width is zero.  This was crashing on width zero because the repeat block
counter was getting set with a very high value.  Using the "View IP"
function in the Java code, which sets the gate widths to 0, was causing
a crash.

Moved debugging code to a separate file.

Signed-off-by: Mike <mike@folsomvillage.com>