Changed interfaceFound to processingFlags1 such that each bit can be used as a flag by the various processing functions rather than the entire word being used as a flag just for the interface found status.
Added code to set/read the IFACE_FOUND flag in processingFlags1 as
required by the change above.

Replaced hardcoded constants with HIT_COUNT_MET and MISS_COUNT_MET as they
should be.

Set GATE_EXCEEDED flag if the signal was found to exceed the gate while
looking for a gate crossing.

Signed-off-by: Mike <mike@folsomvillage.com>