AScan capture can now be free-run or triggered by the signal exceeding a gate.  This allows signals to be captured and displayed clearly.
AScan datasets in free-run mode are now only captured after a request for
an AScan set from the host.  Before, they were being captured with every
shot -- a huge waste of time as most sets were thrown away before the host
requested the next one.

In free-run mode, the host is always receiving the AScan captured just
after the last host request but this is not noticeable.

Various constants in the code were replaced with their named .equ values
to make for better code.

Signed-off-by: Mike Schoonover <mike@folsomvillage.com>