Work in progress -- does  not work!
The GATE_EXCEEDED flag for the results was 0x0004, but this is where the
GATE_MAX_MIN flag should be located according to all other code. Since
GATE_EXCEEDED is in the results flags, it would have been overwriting
the gate's MIN/MAX value which is copied there from the gate's function
flags. It may or may not have been causing problems at times.

The wall code was erroneously hardcoded to use flag 0x002 (MISS_COUNT_MET)
to determine MIN/MAX gate status and that flag has always been zero for
actual field use so the gates defaulted to MAX gates which is the only
mode that was being used anyway. The code has been switched to use the
GATE_MIN_MAX flag, but this causes problems because the host sets the
first wall gate to a MAX and the second to a MIN as the two gates are
used to store the min/max values. The DSP needs both gates to be MAX so
code was added to force them to such.

GATE_MIN_MAX definition has now been dupicated to the gate results section.
Since it is already defined in the gate function section, the two must
match, which is the way the program is designed.

The comments in the findGatePeak function regarding the peak capture have
been mangled in this version -- and are wrong. Need to be fixed.


Signed-off-by: Mike Schoonover <mike@folsomvillage.com>