Fixed problem with comparing trackcount with previous value. Was storing the new value in the "previous" variable before subtracting the two -- result was that the two were always the same.
This code was tested and basically worked except there were only 250
UT shots between TDC control codes -- should have been 500 or so. Might
have been due to gates not being triggered as they were simply put down
into the noise so the signal would exceed the gate and cause the wall to
be processed.

Need to change so that wall reading is stored in map even when gates are
not exceeded -- value from the previous shot should be used.

Signed-off-by: Mike Schoonover <mike@folsomvillage.com>