DMA2 interrupts are now disabled in the DMA2 int handler instead of the timer int handler.
Removed loading DP with #Variables1 from sendPacket so registers accessed
properly.

Fixed code in processAScan where processAScanFast was always being called
after the calls to fast or slow process. This was causing a double hit
for fast modes and an unwanted hit after performing a slow scan.

NOTE: the slow scan method may actually work now...it was abandoned after
it exhibited wonky behavior.

Added ret to end of setupGatesDACs function...was always executing
disableSerialTransmitter before returning, which didn't cause any problems
until changes which have been made to function of the latter.

All interrupts are now disabled on start up -- the IMR register is not
cleared by a reset.

Fixed code which was setting the Stack Pointer -- WAS BEING SET WRONG FOR
A LONG TIME!!!

Moved call to initialize the debugger to the beginning of main so that
any debugger use will be ready.

Signed-off-by: Mike Schoonover <mike@skoonie.com>