Code added to prevent serial port processing while the previous packet is still being handled. NOTE: temporarily changed to always call as a test to solve the packet not returning problem -- did not solve the problem.
Pending interrupts are now cleared before enabling interrupts at the
beginning of main. The IFR flag is NOT cleared by a reset.

Moved code which enables the DMA2 interrupt to the point before where the
DMA might actually be invoked...probably wasn't an issue except where a
breakpoint was inserted between the two. The DMA was interrupting during
the breakpoint and then it was getting cleared as the interrupt enable
code followed the breakpoint.

Moved code which starts the serial port to sendPacket function. Where it
originally was could get called without actually calling sendPacket. The
transmitter would have been turned off in the original code because the
disable code was called periodically, but the new code wouldn't turn it
off in that case because the interrupts which do that are not started
unless sendPacket is called.

Changed timer prescaler from 1 to 3 for test. Changed timer period value
to 60,0000 from 40,000 for test. Still get errors where packet is not
returned in time.

Changed DMA2 to interrupt on frame & block instead of just block...did
not solve problem either.

Commented out all breakpoints.

Added pshm/popm code to the interrupt routines for appropriate registers.
This solved crashing problem.

Switched the debug breakpoint catch routine storeRegistersAndHalt to an
interrupt routine (invoked with "intr 2") so that interrupts are disabled
during a breakpoint to avoid recursive calls back to the function by
interrupt routines containing breakpoints.

Signed-off-by: Mike Schoonover <mike@skoonie.com>